collector/requirements.txt
playwright>=1.49.0
beautifulsoup4>=4.12.0
aiohttp>=3.10.0
python-dotenv>=1.0.0
supabase>=2.3.0
apscheduler>=3.10.0
pillow>=10.1.0
imagehash>=4.3.0
pytest>=7.4.0
python-telegram-bot>=20.0
google-generativeai>=0.3.0
